added search pages and navigation routings
General search - searching for users and groups in the main home page

userSearch - search for users in SelectedGroups pages
